Update README.md
I found your readme this morning and followed it, Unfortunately, the "git clone https://github.com/udacity/CarND-Term1-Starter-Kit.git" command did not download the environment.yml file, so I could not execute the junyper notebook when it came to the thensorflow part. So I figured out that there is no carnd-term1 environment in my python environments, which was because I did not have the environment.yml file. 
For me as a beginner, it was hard to figure out the problem, and I can imagine there are many interested people who would not be able to figure out the problems themselves. Consequently, I propose to expand the readme file.