windows instructions so meta.yml gets run
windows users benefit from meta.yml being run. Renaming this file causes the file to be run automatically.
I've updated the instructions  to windows users to do so.

meta.yml sets a conda variable to a value that does not (sometimes) cause erroneous errors on windows machines that result in various package installs and updates to sometimes fail.   
meta.yml is run automatically when present.   
The value in the meta.yml file was the default value for all conda builds before the 4.2 update.    

The current, conda 4.2, default value should not be used on windows machines, as windows 
1) does not need to check this value-it does not apply to windows machines, and 
2) it misinterprets and misapplies the results of its tests on windows to sometimes erroneously produce  "PaddingErrors" that don't actually exist. 
Padding Errors halt the installation/update of packages whenever they occur.

Seemingly random users will experience this PaddingError, on various packages, because of the nature of the check. (something to do with the length of "placeholders" that various developers sometimes insert into their packages, compared to the length of a pathname on a users machine).  This check should not even be happening for windows, and conda handles the result of this check incorrectly anyway.  Hence the errors.

Since this was the default value for conda < 4.2, it probably does not hurt to have this file run on non-windows systems.
However, since conda is only handling the current value incorrectly on windows, it's better to correct for the specific problem, rather than potentially introduce a future potential side effect for mac or linux users.  It is a windows specific issue.