rename file to meta_windows_patch.yml
windows users need to run this file.
when it is renamed to meta.yml, then it will automatically be run when the code for creating their environment is run.
This file corrects an incorrect default variable setting for conda 4.2 and up.  Prior versions of conda used this setting as the default already, so harm is done if user is running a prior version of conda.

It may also be fine for linux and macos to use this file as well (it was the default prior to conda 4.2), but the issues it causes is unique to windows, so this is an issue for windows users specifically.