Patch for Windows Users (#54)
* Patch for Windows Users

Conda v 4.2 and up changed a default settings which wreaks havoc on many Windows machines.
This file restores the default setting of prior releases, to allow for proper install of many packages to windows.

This file is automatically run first in the build process, in order to allow for customized environment settings.

This file probably should not be run on non-windows machines.
Though, since it used to be the default setting, it might be fine to run it on linux or macos as well..

* windows instructions so meta.yml gets run

windows users benefit from meta.yml being run. Renaming this file causes the file to be run automatically.
I've updated the instructions  to windows users to do so.

meta.yml sets a conda variable to a value that does not (sometimes) cause erroneous errors on windows machines that result in various package installs and updates to sometimes fail.   
meta.yml is run automatically when present.   
The value in the meta.yml file was the default value for all conda builds before the 4.2 update.    

The current, conda 4.2, default value should not be used on windows machines, as windows 
1) does not need to check this value-it does not apply to windows machines, and 
2) it misinterprets and misapplies the results of its tests on windows to sometimes erroneously produce  "PaddingErrors" that don't actually exist. 
Padding Errors halt the installation/update of packages whenever they occur.

Seemingly random users will experience this PaddingError, on various packages, because of the nature of the check. (something to do with the length of "placeholders" that various developers sometimes insert into their packages, compared to the length of a pathname on a users machine).  This check should not even be happening for windows, and conda handles the result of this check incorrectly anyway.  Hence the errors.

Since this was the default value for conda < 4.2, it probably does not hurt to have this file run on non-windows systems.
However, since conda is only handling the current value incorrectly on windows, it's better to correct for the specific problem, rather than potentially introduce a future potential side effect for mac or linux users.  It is a windows specific issue.

* rename meta.yml so it's only run by windows users

windows users should rename this file to meta.yml so it is automatically run and applies the appropriate setting on windows machines.
non-windows users can ignore or delete this file.

* rename file to meta_windows_patch.yml

windows users need to run this file.
when it is renamed to meta.yml, then it will automatically be run when the code for creating their environment is run.
This file corrects an incorrect default variable setting for conda 4.2 and up.  Prior versions of conda used this setting as the default already, so harm is done if user is running a prior version of conda.

It may also be fine for linux and macos to use this file as well (it was the default prior to conda 4.2), but the issues it causes is unique to windows, so this is an issue for windows users specifically.

* fix typo